+++ app/src/main/java/com/example/tanorami/create_build_hero/data/model/BuildHeroFromUser.kt
@@ -8,7 +8,7 @@ data class BuildHeroFromUser(
     val idDecoration: Int,
     val statsEquipment: Array<String>,
     val nickname: String = "",
-    val idBuild: Int? = null,
+    val idBuild: Long? = null,
 ) {
     override fun equals(other: Any?): Boolean {
         if (this === other) return true
