+++ app/src/main/java/com/example/tanorami/viewing_users_build/data/ViewingUsersBuildRepository.kt
@@ -4,5 +4,5 @@ import com.example.tanorami.data.NetworkResult
 import com.example.tanorami.viewing_users_build.data.model.FullBuildHeroFromUser
 
 interface ViewingUsersBuildRepository {
-    suspend fun getHeroBuild(idBuild: Int): NetworkResult<FullBuildHeroFromUser>
+    suspend fun getHeroBuild(idBuild: Long): NetworkResult<FullBuildHeroFromUser>
 }
